:wrench: Introduce slimmed down docker-compose and .env
Currently, I only need nginx, php-fpm, mysql, php-worker, redis, elasticsearch
and the workspace containers. Everything that is not necessary for the correct
functioning of those containers has been stripped out.

Either the repository is to live next to the laravel application under src/ or
to be used as a multi project setup which is primarily useful in development.

The code is to be served from the current/ directory as we will be depending on
Laravel Envoy to handle zero downtime deployment.